+++ App.tsx
@@ -1,18 +1,8 @@
 import * as React from 'react';
-import {NavigationContainer} from '@react-navigation/native';
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import Home from './pages/Home';
-
-const Stack = createNativeStackNavigator();
+import {Router} from './common/routing/Router';
 
 function App(): React.JSX.Element {
-  return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen name="Home" component={Home} />
-      </Stack.Navigator>
-    </NavigationContainer>
-  );
+  return <Router />;
 }
 
 export default App;
