+++ common/Home.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import {View, Text, Button} from 'react-native';
+
+const Home = ({navigation}: any) => {
+  return (
+    <View>
+      <Text>Home Screen</Text>
+      <Button
+        title="Go to Details"
+        onPress={() => navigation.navigate('Details')}
+      />
+    </View>
+  );
+};
+
+export default Home;
