+++ common/routing/Router.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import Home from '../Home';
+import {Details} from '../Details';
+
+const Stack = createNativeStackNavigator();
+
+export const Router = () => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator screenOptions={{headerShown: false}}>
+        <Stack.Screen
+          name="Home"
+          component={Home}
+          options={{title: 'Overview'}}
+        />
+        <Stack.Screen
+          name="Details"
+          component={Details}
+          options={{title: 'Overview'}}
+        />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
