--- pages/Home.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import {View, Text} from 'react-native';
-
-const Home = () => {
-  return (
-    <View>
-      <Text>Home Screen</Text>
-    </View>
-  );
-};
-
-export default Home;
